More fixes and higher warning level

+++ libs/edge/include/edge/hysteresis.hpp
@@ -74,31 +74,31 @@ inline void hysteresis_map(cv::Mat& map, const cv::Mat& dmap, IndexVector& edgel
     }
     if (m[1] > -1) {
       m[1] = -1;
-      edgels.push_back(i + 1);
+      edgels.push_back(static_cast<size_t>(i + 1));
     }
     if (m[-mapstep - 1] > -1) {
       m[-mapstep - 1] = -1;
-      edgels.push_back(i - mapstep - 1);
+      edgels.push_back(static_cast<size_t>(static_cast<ptrdiff_t>(i) - mapstep - 1));
     }
     if (m[-mapstep] > -1) {
       m[-mapstep] = -1;
-      edgels.push_back(i - mapstep);
+      edgels.push_back(static_cast<size_t>(static_cast<ptrdiff_t>(i) - mapstep));
     }
     if (m[-mapstep + 1] > -1) {
       m[-mapstep + 1] = -1;
-      edgels.push_back(i - mapstep + 1);
+      edgels.push_back(static_cast<size_t>(static_cast<ptrdiff_t>(i) - mapstep + 1));
     }
     if (m[mapstep - 1] > -1) {
       m[mapstep - 1] = -1;
-      edgels.push_back(i + mapstep - 1);
+      edgels.push_back(static_cast<size_t>(static_cast<ptrdiff_t>(i) + mapstep - 1));
     }
     if (m[mapstep] > -1) {
       m[mapstep] = -1;
-      edgels.push_back(i + mapstep);
+      edgels.push_back(static_cast<size_t>(static_cast<ptrdiff_t>(i) + mapstep));
     }
     if (m[mapstep + 1] > -1) {
       m[mapstep + 1] = -1;
-      edgels.push_back(i + mapstep + 1);
+      edgels.push_back(static_cast<size_t>(static_cast<ptrdiff_t>(i) + mapstep + 1));
     }
   }
 }
@@ -149,27 +149,27 @@ inline cv::Mat hysteresis_binary(const cv::Mat& dmap, IndexVector& edgels, uchar
     }
     if (m[-mapstep - 1]) {
       m[-mapstep - 1] = 0;
-      edgels.push_back(i - mapstep - 1);
+      edgels.push_back(static_cast<size_t>(i - mapstep - 1));
     }
     if (m[-mapstep]) {
       m[-mapstep] = 0;
-      edgels.push_back(i - mapstep);
+      edgels.push_back(static_cast<size_t>(i - mapstep));
     }
     if (m[-mapstep + 1]) {
       m[-mapstep + 1] = 0;
-      edgels.push_back(i - mapstep + 1);
+      edgels.push_back(static_cast<size_t>(i - mapstep + 1));
     }
     if (m[mapstep - 1]) {
       m[mapstep - 1] = 0;
-      edgels.push_back(i + mapstep - 1);
+      edgels.push_back(static_cast<size_t>(i + mapstep - 1));
     }
     if (m[mapstep]) {
       m[mapstep] = 0;
-      edgels.push_back(i + mapstep);
+      edgels.push_back(static_cast<size_t>(i + mapstep));
     }
     if (m[mapstep + 1]) {
       m[mapstep + 1] = 0;
-      edgels.push_back(i + mapstep + 1);
+      edgels.push_back(static_cast<size_t>(i + mapstep + 1));
     }
   }
   map.setTo(0);
